requirements.txt
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.4
decorator==4.3.2
kiwisolver==1.0.1
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
PyWavelets==1.0.2
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
toolz==0.9.0
